+++ authority-portal-frontend/src/app/pages/registration-process-wizard/registration-process-wizard.module.ts
@@ -9,8 +9,8 @@ import {DevUtilsModule} from 'src/app/common/components/dev-utils/dev-utils.modu
 import {OrganizationCreatePageComponent} from './organization-create-page/organization-create-page.component';
 import {OrganizationCreatePageStateImpl} from './organization-create-page/state/organization-create-page-state-impl';
 import {OrganizationPendingPageComponent} from './organization-pending-page/organization-pending-page.component';
+import {OrganizationRejectedPageComponent} from './organization-rejected-page/organization-rejected-page.component';
 import {RegistrationProcessWizardComponent} from './registration-process-wizard.component';
-import { OrganizationRejectedPageComponent } from './organization-rejected-page/organization-rejected-page.component';
 
 @NgModule({
   imports: [